feat(db/deletion) Support database resource deletion when the resource was deleted first through Scalingo API
