+++ internal/usecases/database/base/manager_test.go
@@ -1,7 +1,6 @@
 package database
 
 import (
-	"errors"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -32,208 +31,6 @@ func TestNewManager(t *testing.T) {
 
 		require.EqualError(t, err, "empty api token")
 		require.Nil(t, dbManager)
-	})
-}
-
-func TestManager_CreateDatabase(t *testing.T) {
-	t.Run("it successfully creates database", func(t *testing.T) {
-		// Given
-		ctx := t.Context()
-		ctrl := gomock.NewController(t)
-		scClient := scalingomock.NewMockClient(ctrl)
-
-		manager := manager{
-			scClient: scClient,
-		}
-
-		dbRequested := domain.Database{
-			Name: "PG test",
-			Type: domain.DatabaseTypePostgreSQL,
-			Plan: "postgresql-dr-enterprise-4096",
-		}
-
-		dbCreated := dbRequested
-		dbCreated.ID = databaseID
-
-		scClient.EXPECT().CreateDatabase(ctx, dbRequested).Return(dbCreated, nil)
-
-		// When
-		res, err := manager.CreateDatabase(ctx, dbRequested)
-
-		// Then
-		require.NoError(t, err)
-		require.Equal(t, dbCreated, res)
-	})
-}
-
-func TestManager_GetDatabase(t *testing.T) {
-	t.Run("it fails because of empty ID", func(t *testing.T) {
-		ctx := t.Context()
-		manager := manager{}
-		res, err := manager.GetDatabase(ctx, "")
-
-		require.EqualError(t, err, "empty database id")
-		require.Empty(t, res)
-	})
-
-	t.Run("it successfully gets database", func(t *testing.T) {
-		// Given
-		ctx := t.Context()
-		ctrl := gomock.NewController(t)
-		scClient := scalingomock.NewMockClient(ctrl)
-
-		manager := manager{
-			scClient: scClient,
-		}
-
-		db := domain.Database{
-			ID:      databaseID,
-			AddonID: addonID,
-			Name:    "PG test",
-			Type:    domain.DatabaseTypePostgreSQL,
-			Plan:    "postgresql-dr-enterprise-4096",
-		}
-
-		scClient.EXPECT().GetDatabase(ctx, databaseID).Return(db, nil)
-		scClient.EXPECT().ListFirewallRules(ctx, databaseID, addonID)
-
-		// When
-		res, err := manager.GetDatabase(ctx, databaseID)
-
-		// Then
-		require.NoError(t, err)
-		require.Equal(t, db, res)
-	})
-}
-
-func TestManager_GetDatabaseURL(t *testing.T) {
-	t.Run("it fails because of invalid datatype", func(t *testing.T) {
-		ctx := t.Context()
-		manager := manager{}
-		db := domain.Database{
-			Type: "invalid_type",
-		}
-		res, err := manager.GetDatabaseURL(ctx, db)
-
-		require.EqualError(t, err, `to database type name: no matching type for "invalid_type"`)
-		require.Empty(t, res)
-	})
-
-	t.Run("it successfully gets database URL", func(t *testing.T) {
-		// Given
-		ctx := t.Context()
-		ctrl := gomock.NewController(t)
-		scClient := scalingomock.NewMockClient(ctrl)
-
-		manager := manager{
-			scClient: scClient,
-		}
-
-		db := domain.Database{
-			ID:    databaseID,
-			AppID: appID,
-			Type:  domain.DatabaseTypePostgreSQL,
-		}
-
-		dbURL := domain.DatabaseURL{
-			Name:  "SCALINGO_POSTGRESQL_URL",
-			Value: "postgresql_connection_string",
-		}
-
-		scClient.EXPECT().FindApplicationVariable(ctx, db.AppID, "SCALINGO_POSTGRESQL_URL").Return(dbURL.Value, nil)
-
-		// When
-		res, err := manager.GetDatabaseURL(ctx, db)
-
-		// Then
-		require.NoError(t, err)
-		require.Equal(t, dbURL, res)
-	})
-}
-
-func TestManager_UpdateDatabase(t *testing.T) {
-	t.Run("it fails when getting database", func(t *testing.T) {
-		// Given
-		ctx := t.Context()
-		ctrl := gomock.NewController(t)
-		scClient := scalingomock.NewMockClient(ctrl)
-
-		manager := manager{
-			scClient: scClient,
-		}
-
-		scClient.EXPECT().GetDatabase(ctx, databaseID).Return(domain.Database{}, errors.New("boom"))
-
-		// When
-		status, err := manager.UpdateDatabase(ctx, databaseID, domain.Database{})
-
-		// Then
-		require.Equal(t, domain.DatabaseStatusUnknown, status)
-		require.ErrorContains(t, err, "get database")
-	})
-
-	t.Run("it updates firewall rules only when database is provisioning", func(t *testing.T) {
-		// Given
-		ctx := t.Context()
-		ctrl := gomock.NewController(t)
-		scClient := scalingomock.NewMockClient(ctrl)
-
-		manager := manager{
-			scClient: scClient,
-		}
-
-		currentDB := domain.Database{
-			ID:      databaseID,
-			AddonID: addonID,
-			Plan:    "postgresql-dr-starter-4096",
-			Status:  domain.DatabaseStatusProvisioning,
-		}
-		expectedDB := domain.Database{
-			Plan: "postgresql-dr-enterprise-4096",
-		}
-
-		scClient.EXPECT().GetDatabase(ctx, databaseID).Return(currentDB, nil)
-		scClient.EXPECT().ListFirewallRules(ctx, databaseID, addonID).Return(nil, nil)
-
-		// When
-		status, err := manager.UpdateDatabase(ctx, databaseID, expectedDB)
-
-		// Then
-		require.NoError(t, err)
-		require.Equal(t, domain.DatabaseStatusProvisioning, status)
-	})
-
-	t.Run("it updates plan when database is running", func(t *testing.T) {
-		// Given
-		ctx := t.Context()
-		ctrl := gomock.NewController(t)
-		scClient := scalingomock.NewMockClient(ctrl)
-
-		manager := manager{
-			scClient: scClient,
-		}
-
-		currentDB := domain.Database{
-			ID:      databaseID,
-			AddonID: addonID,
-			Plan:    "postgresql-dr-enterprise-2048",
-			Status:  domain.DatabaseStatusRunning,
-		}
-		expectedDB := domain.Database{
-			Plan: "postgresql-dr-enterprise-4096",
-		}
-
-		scClient.EXPECT().GetDatabase(ctx, databaseID).Return(currentDB, nil)
-		scClient.EXPECT().ListFirewallRules(ctx, databaseID, addonID).Return(nil, nil)
-		scClient.EXPECT().UpdateDatabasePlan(ctx, currentDB, expectedDB.Plan).
-			Return(domain.DatabaseStatusProvisioning, nil)
-
-		// When
-		status, err := manager.UpdateDatabase(ctx, databaseID, expectedDB)
-
-		// Then
-		require.NoError(t, err)
-		require.Equal(t, domain.DatabaseStatusProvisioning, status)
 	})
 }
 
@@ -295,35 +92,6 @@ func TestManager_updateDatabaseWithProvisioning(t *testing.T) {
 		require.Equal(t, domain.DatabaseStatusProvisioning, status)
 	})
 }
-
-func TestManager_DeleteDatabase(t *testing.T) {
-	t.Run("it fails because of empty ID", func(t *testing.T) {
-		ctx := t.Context()
-		manager := manager{}
-		err := manager.DeleteDatabase(ctx, "")
-
-		require.EqualError(t, err, "empty database id")
-	})
-
-	t.Run("it successfully deletes database", func(t *testing.T) {
-		// Given
-		ctx := t.Context()
-		ctrl := gomock.NewController(t)
-		scClient := scalingomock.NewMockClient(ctrl)
-
-		manager := manager{
-			scClient: scClient,
-		}
-
-		scClient.EXPECT().DeleteDatabase(ctx, databaseID).Return(nil)
-
-		// When
-		err := manager.DeleteDatabase(ctx, databaseID)
-		// Then
-		require.NoError(t, err)
-	})
-}
-
 func TestToDatabaseTypeName(t *testing.T) {
 	ctx := t.Context()
